Infinity should be treated as plural
In plural.js, `parseInt()`[0] is used to check if the given input is a
number. However the radix argument for `parseInt()` is omitted which can
produce unexpected behaviors; this is discouraged by the MDN[1].

Anyways if we are using Infinity (for example with `$ngettext(singular, plural,
Infinity)`) then the expected translation should be plural instead of
singular.

[0] https://tc39.es/ecma262/#sec-parseint-string-radix
[1] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/parseInt