translate: add 'language' parameter
It is convenient to fetch a translation for a specific language
without running 'translate.initTranslations()'
to change  the default language.
'language' property in the config could be set
outside of translate.js
but code that imports only 'vue-gettext' may not have access to the config.

Use-case:
vue-gettext is used for all translations in a nodejs server
that accepts language as a request parameter (or query).
It should be able to pass the language to '$gettext()'
if it is not the default language.